refactor: Update logo with props
