+++ src/components/Molucules/Logo/Logo.type.ts
@@ -1,0 +1,5 @@
+export type LogoProps = {
+    title: string;
+    link: string;
+  };
+
